+++ src/DependencyInjection/AjglLogoutRedirectorExtension.php
@@ -19,6 +19,7 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\Security\Http\Event\LogoutEvent;
 
 class AjglLogoutRedirectorExtension extends Extension
 {
@@ -29,19 +30,26 @@ class AjglLogoutRedirectorExtension extends Extension
 
         $config = $this->processConfiguration(new Configuration(), $configs);
 
-        $this->configureLogoutSuccessHandlerDecorators($container, $config['redirectors']);
+        $this->configureLogoutRedirector($container, $config['redirectors']);
     }
 
-    private function configureLogoutSuccessHandlerDecorators(ContainerBuilder $container, array $redirectors): void
+    private function configureLogoutRedirector(ContainerBuilder $container, array $redirectors): void
     {
         foreach ($redirectors as $name => $config) {
             $redirectorChildDefinition = new ChildDefinition('ajgl_logout_redirector.security.logout.redirector');
             $redirectorChildDefinition->replaceArgument(1, $config);
             $container->setDefinition('ajgl_logout_redirector.security.logout.redirector.'.$name, $redirectorChildDefinition);
 
-            $handlerChildDefinition = new ChildDefinition('ajgl_logout_redirector.security.logout.success_handler');
-            $handlerChildDefinition->replaceArgument(0, new Reference('ajgl_logout_redirector.security.logout.redirector.'.$name));
-            $container->setDefinition('ajgl_logout_redirector.security.logout.success_handler.'.$name, $handlerChildDefinition);
+            if (class_exists(LogoutEvent::class)) {
+                $listenerChildDefinition = new ChildDefinition('ajgl_logout_redirector.security.logout.listener');
+                $listenerChildDefinition->replaceArgument(0, new Reference('ajgl_logout_redirector.security.logout.redirector.'.$name));
+                $listenerChildDefinition->addTag('kernel.event_subscriber');
+                $container->setDefinition('ajgl_logout_redirector.security.logout.listener.'.$name, $listenerChildDefinition);
+            } else {
+                $handlerChildDefinition = new ChildDefinition('ajgl_logout_redirector.security.logout.success_handler');
+                $handlerChildDefinition->replaceArgument(0, new Reference('ajgl_logout_redirector.security.logout.redirector.'.$name));
+                $container->setDefinition('ajgl_logout_redirector.security.logout.success_handler.'.$name, $handlerChildDefinition);
+            }
         }
     }
 }