+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 * Add new `LogoutRedirector` class
+* Add new `LogoutRedirectorEventListener` to support new `LogoutEvent`
 
 ### Changed